Fix wrong definition of floating-point registers.

+++ src/cpu/interpreter/interpreter_pairedsingle.cpp
@@ -7,12 +7,12 @@
 static void
 ps_abs(ThreadState *state, Instruction instr)
 {
-   double b0, b1, d0, d1;
+   float b0, b1, d0, d1;
    b0 = state->fpr[instr.frB].paired0;
    b1 = state->fpr[instr.frB].paired1;
 
-   d1 = std::fabs(b1);
-   d0 = std::fabs(b0);
+   d1 = std::fabsf(b1);
+   d0 = std::fabsf(b0);
 
    state->fpr[instr.frD].paired0 = d0;
    state->fpr[instr.frD].paired1 = d1;
@@ -26,7 +26,7 @@ ps_abs(ThreadState *state, Instruction instr)
 static void
 ps_add(ThreadState *state, Instruction instr)
 {
-   double a0, a1, b0, b1, d0, d1;
+   float a0, a1, b0, b1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
    a1 = state->fpr[instr.frA].paired1;
@@ -59,7 +59,7 @@ ps_add(ThreadState *state, Instruction instr)
 static void
 ps_div(ThreadState *state, Instruction instr)
 {
-   double a0, a1, b0, b1, d0, d1;
+   float a0, a1, b0, b1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
    a1 = state->fpr[instr.frA].paired1;
@@ -105,7 +105,7 @@ template<unsigned flags>
 static void
 maddGeneric(ThreadState *state, Instruction instr)
 {
-   double a0, a1, b0, b1, c0, c1, d0, d1;
+   float a0, a1, b0, b1, c0, c1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
    a1 = state->fpr[instr.frA].paired1;
@@ -193,7 +193,7 @@ template<unsigned flags = 0>
 static void
 mergeGeneric(ThreadState *state, Instruction instr)
 {
-   double d0, d1;
+   float d0, d1;
 
    if (flags & MergeValue0) {
       d0 = state->fpr[instr.frA].paired1;
@@ -261,7 +261,7 @@ template<unsigned flags = 0>
 static void
 msubGeneric(ThreadState *state, Instruction instr)
 {
-   double a0, a1, b0, b1, c0, c1, d0, d1;
+   float a0, a1, b0, b1, c0, c1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
    a1 = state->fpr[instr.frA].paired1;
@@ -330,7 +330,7 @@ template<unsigned flags>
 static void
 mulGeneric(ThreadState *state, Instruction instr)
 {
-   double a0, a1, c0, c1, d0, d1;
+   float a0, a1, c0, c1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
    a1 = state->fpr[instr.frA].paired1;
@@ -391,13 +391,13 @@ ps_muls1(ThreadState *state, Instruction instr)
 static void
 ps_nabs(ThreadState *state, Instruction instr)
 {
-   double b0, b1, d0, d1;
+   float b0, b1, d0, d1;
 
    b0 = state->fpr[instr.frB].paired0;
    b1 = state->fpr[instr.frB].paired1;
 
-   d1 = -std::fabs(b1);
-   d0 = -std::fabs(b0);
+   d1 = -std::fabsf(b1);
+   d0 = -std::fabsf(b0);
 
    state->fpr[instr.frD].paired0 = d0;
    state->fpr[instr.frD].paired1 = d1;
@@ -411,7 +411,7 @@ ps_nabs(ThreadState *state, Instruction instr)
 static void
 ps_neg(ThreadState *state, Instruction instr)
 {
-   double b0, b1, d0, d1;
+   float b0, b1, d0, d1;
 
    b0 = state->fpr[instr.frB].paired0;
    b1 = state->fpr[instr.frB].paired1;
@@ -431,7 +431,7 @@ ps_neg(ThreadState *state, Instruction instr)
 static void
 ps_res(ThreadState *state, Instruction instr)
 {
-   double b0, b1, d0, d1;
+   float b0, b1, d0, d1;
 
    b0 = state->fpr[instr.frB].paired0;
    b1 = state->fpr[instr.frB].paired1;
@@ -459,7 +459,7 @@ ps_res(ThreadState *state, Instruction instr)
 static void
 ps_rsqrte(ThreadState *state, Instruction instr)
 {
-   double b0, b1, d0, d1;
+   float b0, b1, d0, d1;
 
    b0 = state->fpr[instr.frB].paired0;
    b1 = state->fpr[instr.frB].paired1;
@@ -470,8 +470,8 @@ ps_rsqrte(ThreadState *state, Instruction instr)
    state->fpscr.vxsqrt |=
       is_negative(b0) || is_negative(b1);
 
-   d1 = 1.0f / std::sqrt(b1);
-   d0 = 1.0f / std::sqrt(b0);
+   d1 = 1.0f / std::sqrtf(b1);
+   d0 = 1.0f / std::sqrtf(b0);
    updateFPSCR(state);
    updateFPRF(state, d0);
 
@@ -487,7 +487,7 @@ ps_rsqrte(ThreadState *state, Instruction instr)
 static void
 ps_sel(ThreadState *state, Instruction instr)
 {
-   double a0, a1, b0, b1, c0, c1, d0, d1;
+   float a0, a1, b0, b1, c0, c1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
    a1 = state->fpr[instr.frA].paired1;
@@ -513,7 +513,7 @@ ps_sel(ThreadState *state, Instruction instr)
 static void
 ps_sub(ThreadState *state, Instruction instr)
 {
-   double a0, a1, b0, b1, d0, d1;
+   float a0, a1, b0, b1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
    a1 = state->fpr[instr.frA].paired1;
@@ -553,7 +553,7 @@ template<unsigned flags = 0>
 static void
 sumGeneric(ThreadState *state, Instruction instr)
 {
-   double a0, b0, b1, c0, c1, d0, d1;
+   float a0, b0, b1, c0, c1, d0, d1;
 
    a0 = state->fpr[instr.frA].paired0;
